+++ pallets/staking/src/benchmarking.rs
@@ -21,6 +21,7 @@ use frame_support::{
     assert_ok, ensure,
     sp_runtime::traits::StaticLookup,
     traits::{Currency, Get},
+    BoundedVec,
 };
 use frame_system::{EventRecord, RawOrigin};
 use pallet_parameters::{SignersInfo, SignersSize};
@@ -77,8 +78,12 @@ fn prep_bond_and_validate<T: Config>(
         reward_destination,
     ));
 
-    let server_info =
-        ServerInfo { tss_account: threshold, x25519_public_key, endpoint: vec![20, 20] };
+    let server_info = ServerInfo {
+        tss_account: threshold,
+        x25519_public_key,
+        endpoint: vec![20, 20],
+        provisioning_certification_key: BoundedVec::with_max_capacity(),
+    };
 
     if validate_also {
         assert_ok!(<Staking<T>>::validate(
@@ -120,6 +125,7 @@ benchmarks! {
       endpoint: vec![20, 20],
       tss_account: _bonder.clone(),
       x25519_public_key: NULL_ARR,
+      provisioning_certification_key: BoundedVec::with_max_capacity(),
     };
     assert_last_event::<T>(Event::<T>::ThresholdAccountChanged(bonder, server_info).into());
   }
@@ -161,6 +167,7 @@ benchmarks! {
         tss_account: threshold.clone(),
         x25519_public_key: NULL_ARR,
         endpoint: vec![20],
+        provisioning_certification_key: BoundedVec::with_max_capacity(),
     };
 
   }:  _(RawOrigin::Signed(bonder.clone()), validator_preference, server_info)